requirements.txt
discord.py
rich
dotenv
watchdog
Pillow
aiohttp
sqlalchemy
aiosqlite
greenlet